# backend/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3001",
  "basePath": "/api",
  "tags": [
    {
      "name": "post",
      "description": "post router"
    },
    {
      "name": "user",
      "description": "user router"
    },
    {
      "name": "category",
      "description": "category router"
    },
    {
      "name": "test",
      "description": "test router"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/category/category/{id}": {
      "delete": {
        "tags": [
          "category"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "category not found"
          }
        }
      },
      "get": {
        "tags": [
          "category"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/category/category": {
      "get": {
        "tags": [
          "category"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/user/user": {
      "post": {
        "tags": [
          "user"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "get": {
        "tags": [
          "user"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/user/user/{id}": {
      "get": {
        "tags": [
          "user"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/post/drafts": {
      "get": {
        "tags": [
          "post"
        ],
        "summary": "Get all drafts",
        "description": "Some description...",
        "operationId": "Your_operationId_here",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/post/feed": {
      "get": {
        "tags": [
          "post"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/post/filterPosts": {
      "get": {
        "tags": [
          "post"
        ],
        "description": "",
        "parameters": [
          {
            "name": "searchString",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/post/post": {
      "post": {
        "tags": [
          "post"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Some description...",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "example": "title here"
                },
                "content": {
                  "type": "string",
                  "example": "content here"
                },
                "authorEmail": {
                  "type": "string",
                  "example": "email here"
                },
                "catergoryId": {
                  "type": "string",
                  "example": "catid here"
                }
              },
              "required": [
                "title",
                "content",
                "authorEmail",
                "catergoryId"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/post/post/{id}": {
      "delete": {
        "tags": [
          "post"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "post not found"
          }
        }
      },
      "get": {
        "tags": [
          "post"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/post/publish/{id}": {
      "put": {
        "tags": [
          "post"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}